Watch_Tower/requirements.txt
dj-database-url==0.4.1
Django==1.9.6
django-bootstrap3==7.0.1
django-filter==0.13.0
djangorestframework==3.3.3
gunicorn==19.6.0
Markdown==2.6.6
Pillow==3.2.0
psycopg2==2.6.1
python-dateutil==2.5.3
six==1.10.0
whitenoise==3.2
